refactor: rename base base document ot base doc
