+++ docarray/array/abstract_array.py
@@ -19,7 +19,7 @@ from typing import (
 
 import numpy as np
 
-from docarray.base_document import BaseDoc
+from docarray.base_doc import BaseDoc
 from docarray.display.document_array_summary import DocArraySummary
 from docarray.typing.abstract_type import AbstractType
 from docarray.utils._typing import change_cls_name
