+++ docarray/base_doc/__init__.py
@@ -1,0 +1,6 @@
+from docarray.base_doc.any_doc import AnyDoc
+from docarray.base_doc.base_node import BaseNode
+from docarray.base_doc.doc import BaseDoc
+from docarray.base_doc.doc_response import DocResponse
+
+__all__ = ['AnyDoc', 'BaseDoc', 'BaseNode', 'DocResponse']
