+++ docarray/base_doc/mixins/__init__.py
@@ -1,0 +1,4 @@
+from docarray.base_doc.mixins.io import IOMixin
+from docarray.base_doc.mixins.update import UpdateMixin
+
+__all__ = ['IOMixin', 'UpdateMixin']
