+++ docarray/base_doc/mixins/io.py
@@ -17,7 +17,7 @@ from typing import (
 import numpy as np
 from typing_inspect import is_union_type
 
-from docarray.base_document.base_node import BaseNode
+from docarray.base_doc.base_node import BaseNode
 from docarray.typing import NdArray
 from docarray.typing.proto_register import _PROTO_TYPE_NAME_TO_CLASS
 from docarray.utils.compress import _compress_bytes, _decompress_bytes
