--- docarray/base_document/__init__.py
@@ -1,6 +1,0 @@
-from docarray.base_document.any_doc import AnyDoc
-from docarray.base_document.base_node import BaseNode
-from docarray.base_document.doc import BaseDoc
-from docarray.base_document.doc_response import DocResponse
-
-__all__ = ['AnyDoc', 'BaseDoc', 'BaseNode', 'DocResponse']
