--- docarray/base_document/mixins/__init__.py
@@ -1,4 +1,0 @@
-from docarray.base_document.mixins.io import IOMixin
-from docarray.base_document.mixins.update import UpdateMixin
-
-__all__ = ['IOMixin', 'UpdateMixin']
