+++ docarray/documents/mesh/vertices_and_faces.py
@@ -1,6 +1,6 @@
 from typing import Any, Type, TypeVar, Union
 
-from docarray.base_document import BaseDoc
+from docarray.base_doc import BaseDoc
 from docarray.typing.tensor.tensor import AnyTensor
 
 T = TypeVar('T', bound='VerticesAndFaces')
