+++ tests/integrations/document/test_to_json.py
@@ -2,8 +2,8 @@ import numpy as np
 import pytest
 import torch
 
-from docarray.base_document import BaseDoc
-from docarray.base_document.io.json import orjson_dumps
+from docarray.base_doc import BaseDoc
+from docarray.base_doc.io.json import orjson_dumps
 from docarray.typing import AnyUrl, NdArray, TorchTensor
