+++ tests/units/typing/tensor/test_tensor.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.base_document.io.json import orjson_dumps
+from docarray.base_doc.io.json import orjson_dumps
 from docarray.typing import AudioNdArray, NdArray, TorchTensor
 from docarray.typing.tensor import NdArrayEmbedding
