+++ tests/units/typing/url/test_any_url.py
@@ -1,7 +1,7 @@
 import pytest
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.base_document.io.json import orjson_dumps
+from docarray.base_doc.io.json import orjson_dumps
 from docarray.typing import AnyUrl
