+++ tests/units/typing/url/test_image_url.py
@@ -6,7 +6,7 @@ import pytest
 from PIL import Image
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.base_document.io.json import orjson_dumps
+from docarray.base_doc.io.json import orjson_dumps
 from docarray.typing import ImageUrl
 
 CUR_DIR = os.path.dirname(os.path.abspath(__file__))
