+++ tests/units/typing/url/test_mesh_url.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.base_document.io.json import orjson_dumps
+from docarray.base_doc.io.json import orjson_dumps
 from docarray.typing import Mesh3DUrl, NdArray
 from tests import TOYDATA_DIR
