+++ tests/units/typing/url/test_text_url.py
@@ -4,9 +4,8 @@ import urllib
 import pytest
 from pydantic import parse_obj_as, schema_json_of
 
-from docarray.base_document.io.json import orjson_dumps
+from docarray.base_doc.io.json import orjson_dumps
 from docarray.typing import TextUrl
-
 from tests import TOYDATA_DIR
 
 REMOTE_TEXT_FILE = 'https://de.wikipedia.org/wiki/Brixen'
